Split circular_queue (single producer) and circular_queue_mp (multi producer) into separate sources.
